changed module name and fixed name collision of built binary and photerm package by renaming photerm package to photerm_src
