+++ .gitignore
@@ -11,3 +11,4 @@
 *.test.exe
 imagestuff
 *.ttf
+photerm
