+++ photerm_src/text_source.go
